rename : Generator 를 util 패키지로 이동
