+++ src/main/java/lotto/utils/LottoNumbersGenerator.java
@@ -1,6 +1,8 @@
-package lotto.domain;
+package lotto.utils;
 
 import java.util.Set;
+
+import lotto.domain.LottoNumber;
 
 @FunctionalInterface
 public interface LottoNumbersGenerator {
